+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Options may now be required.
+
 ### Changed
 - Improve error handling in custom yaml unmarshalling.
 - Exit for unexpected errors occuring during when clause validation.